refactored. code organization
