+++ lib/cheetah.rb
@@ -1,7 +1,7 @@
 require 'cheetah/message'
 require 'cheetah/exception'
 require 'cheetah/messenger/messenger'
-require 'transactional_cheetah'
+require 'cheetah/transactional_cheetah'
 
 Dir["#{File.dirname(__FILE__)}/cheetah/messenger/*.rb"].each {|f| require f}
