+++ lib/cheetah/transactional_cheetah.rb
